chore: move plugin-babel-source to scripts/
