+++ Gulpfile.js
@@ -13,7 +13,7 @@ const path = require("path");
 const rollup = require("rollup");
 const rollupAlias = require("@rollup/plugin-alias");
 const rollupBabel = require("rollup-plugin-babel");
-const rollupBabelSource = require("./rollup-plugin-babel-source");
+const rollupBabelSource = require("./scripts/rollup-plugin-babel-source");
 const rollupCommonJs = require("rollup-plugin-commonjs");
 const rollupJson = require("@rollup/plugin-json");
 const rollupNodeBuiltins = require("rollup-plugin-node-builtins");
